refactor(survey): decompose the report god-file by role (WS-D)
Split internal/wifi/survey/report.go (981 LOC) into four role-based
files in the same package — pure move, no behavior change:

- report.go    PDF layout consts, ReportOptions, ReportGenerator,
               Generate, Manager.GenerateReport (orchestration core)
- sections.go  the add* PDF section builders
- stats.go     survey/floor statistics, channel usage, recommendations
- format.go    label/colour/string formatters + sample extraction

All four are now under the 600-line size-gate threshold. Declaration
sets and code-line multisets verified identical to the original.

Related to #187
