+++ internal/wifi/survey/report.go
@@ -4,9 +4,6 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"sort"
-	"strconv"
-	"time"
 
 	"github.com/go-pdf/fpdf"
 )
@@ -182,791 +179,6 @@ func (g *ReportGenerator) Generate() ([]byte, error) {
 	}
 
 	return buf.Bytes(), nil
-}
-
-// addCoverPage creates the report cover page.
-func (g *ReportGenerator) addCoverPage() {
-	g.pdf.AddPage()
-
-	// Header with company name if provided
-	if g.options.CompanyName != "" {
-		g.pdf.SetFont("Arial", "", pdfFontSizeBody)
-		g.pdf.SetTextColor(pdfColorGrayLight, pdfColorGrayLight, pdfColorGrayLight)
-		g.pdf.CellFormat(0, pdfSpacingSection, g.options.CompanyName, "", 1, "C", false, 0, "")
-	}
-
-	// Main title
-	g.pdf.Ln(pdfSpacingCover)
-	g.pdf.SetFont("Arial", "B", pdfFontSizeTitle)
-	g.pdf.SetTextColor(0, 0, 0)
-	g.pdf.CellFormat(0, pdfSpacingTitle, "WiFi Site Survey Report", "", 1, "C", false, 0, "")
-
-	// Survey name
-	g.pdf.Ln(pdfSpacingSection)
-	g.pdf.SetFont("Arial", "", pdfFontSizeSubtitle)
-	g.pdf.SetTextColor(pdfColorGrayDark, pdfColorGrayDark, pdfColorGrayDark)
-	g.pdf.CellFormat(0, pdfSpacingSection, g.survey.Name, "", 1, "C", false, 0, "")
-
-	// Description if provided
-	if g.survey.Description != "" {
-		g.pdf.Ln(pdfSpacingSmall)
-		g.pdf.SetFont("Arial", "I", pdfFontSizeBody)
-		g.pdf.SetTextColor(pdfColorGrayLight, pdfColorGrayLight, pdfColorGrayLight)
-		g.pdf.MultiCell(0, pdfSpacingMedium, g.survey.Description, "", "C", false)
-	}
-
-	// Date info
-	g.pdf.Ln(pdfSpacingHuge)
-	g.pdf.SetFont("Arial", "", pdfFontSizeBody)
-	g.pdf.SetTextColor(pdfColorGrayMedium, pdfColorGrayMedium, pdfColorGrayMedium)
-
-	dateStr := time.Now().Format("January 2, 2006")
-	g.pdf.CellFormat(0, pdfSpacingLarge, fmt.Sprintf("Report Generated: %s", dateStr), "", 1, "C", false, 0, "")
-
-	surveyDate := g.survey.CreatedAt.Format("January 2, 2006")
-	g.pdf.CellFormat(0, pdfSpacingLarge, fmt.Sprintf("Survey Date: %s", surveyDate), "", 1, "C", false, 0, "")
-
-	// Status
-	g.pdf.Ln(pdfSpacingSmall)
-	g.pdf.SetFont("Arial", "B", pdfFontSizeBody)
-	statusColor := getStatusColor(g.survey.Status)
-	g.pdf.SetTextColor(statusColor[0], statusColor[1], statusColor[2])
-	g.pdf.CellFormat(0, pdfSpacingLarge, fmt.Sprintf("Status: %s", g.survey.Status), "", 1, "C", false, 0, "")
-
-	// Building info
-	g.pdf.Ln(pdfSpacingMajor)
-	g.pdf.SetFont("Arial", "", pdfFontSizeNormal)
-	g.pdf.SetTextColor(pdfColorGrayMedium, pdfColorGrayMedium, pdfColorGrayMedium)
-
-	floorCount := len(g.survey.Floors)
-	sampleCount := len(g.survey.GetAllSamples())
-	g.pdf.CellFormat(
-		0,
-		pdfSpacingNormal,
-		fmt.Sprintf("Floors: %d | Sample Points: %d", floorCount, sampleCount),
-		"",
-		1,
-		"C",
-		false,
-		0,
-		"",
-	)
-}
-
-// addExecutiveSummary adds the executive summary section.
-func (g *ReportGenerator) addExecutiveSummary() {
-	g.pdf.AddPage()
-	g.addSectionHeader("Executive Summary")
-
-	// Calculate overall statistics
-	allSamples := g.survey.GetAllSamples()
-	stats := calculateSurveyStats(allSamples)
-
-	// Coverage score card
-	g.pdf.Ln(pdfSpacingSmall)
-	g.addStatCard(
-		"Overall Coverage Score",
-		fmt.Sprintf("%.0f%%", stats.CoverageScore),
-		getCoverageGrade(stats.CoverageScore),
-	)
-
-	// Key metrics table
-	g.pdf.Ln(pdfSpacingSection)
-	g.pdf.SetFont("Arial", "B", pdfFontSizeBody)
-	g.pdf.CellFormat(0, pdfSpacingLarge, "Key Metrics", "", 1, "L", false, 0, "")
-
-	g.pdf.SetFont("Arial", "", pdfFontSizeSmall)
-	metrics := []struct {
-		label string
-		value string
-	}{
-		{"Total Sample Points", strconv.Itoa(stats.TotalSamples)},
-		{"Average Signal Strength", fmt.Sprintf("%d dBm", stats.AvgRSSI)},
-		{"Minimum Signal", fmt.Sprintf("%d dBm", stats.MinRSSI)},
-		{"Maximum Signal", fmt.Sprintf("%d dBm", stats.MaxRSSI)},
-		{"Weak Coverage Areas", strconv.Itoa(stats.WeakAreas)},
-		{"Dead Zones Detected", strconv.Itoa(stats.DeadZones)},
-	}
-
-	for _, m := range metrics {
-		g.addMetricRow(m.label, m.value)
-	}
-
-	// Signal distribution breakdown
-	g.pdf.Ln(pdfSpacingSection)
-	g.pdf.SetFont("Arial", "B", pdfFontSizeBody)
-	g.pdf.CellFormat(0, pdfSpacingLarge, "Signal Quality Distribution", "", 1, "L", false, 0, "")
-
-	g.pdf.SetFont("Arial", "", pdfFontSizeSmall)
-	signalDist := []struct {
-		label   string
-		percent float64
-		color   []int
-	}{
-		{"Excellent (> -50 dBm)", stats.ExcellentPercent, []int{40, 167, 69}},
-		{"Good (-50 to -65 dBm)", stats.GoodPercent, []int{144, 238, 144}},
-		{"Fair (-65 to -75 dBm)", stats.FairPercent, []int{255, 193, 7}},
-		{"Poor (-75 to -85 dBm)", stats.PoorPercent, []int{255, 128, 0}},
-		{"Dead (< -85 dBm)", stats.DeadPercent, []int{220, 53, 69}},
-	}
-
-	for _, dist := range signalDist {
-		g.addDistributionBar(dist.label, dist.percent, dist.color)
-	}
-}
-
-// addFloorAnalysis adds per-floor analysis sections.
-func (g *ReportGenerator) addFloorAnalysis() {
-	floors := g.survey.Floors
-	if len(floors) == 0 {
-		return
-	}
-
-	// Sort floors by level
-	sortedFloors := make([]*Floor, len(floors))
-	copy(sortedFloors, floors)
-	sort.Slice(sortedFloors, func(i, j int) bool {
-		return sortedFloors[i].Level < sortedFloors[j].Level
-	})
-
-	for _, floor := range sortedFloors {
-		g.addFloorSection(floor)
-	}
-}
-
-// addFloorSection adds analysis for a single floor.
-func (g *ReportGenerator) addFloorSection(floor *Floor) {
-	g.pdf.AddPage()
-	g.addSectionHeader(fmt.Sprintf("Floor Analysis: %s", floor.Name))
-
-	// Floor info
-	g.pdf.SetFont("Arial", "", pdfFontSizeSmall)
-	g.pdf.SetTextColor(pdfColorGrayMedium, pdfColorGrayMedium, pdfColorGrayMedium)
-	g.pdf.CellFormat(
-		0,
-		pdfSpacingMedium,
-		fmt.Sprintf("Level: %d | Samples: %d", floor.Level, len(floor.Samples)),
-		"",
-		1,
-		"L",
-		false,
-		0,
-		"",
-	)
-
-	// Floor plan dimensions if available
-	if floor.FloorPlan != nil {
-		g.pdf.CellFormat(
-			0,
-			pdfSpacingMedium,
-			fmt.Sprintf("Dimensions: %d x %d px", floor.FloorPlan.Width, floor.FloorPlan.Height),
-			"",
-			1,
-			"L",
-			false,
-			0,
-			"",
-		)
-		if floor.FloorPlan.ScaleM > 0 {
-			g.pdf.CellFormat(
-				0,
-				pdfSpacingMedium,
-				fmt.Sprintf("Scale: %.2f m/px", floor.FloorPlan.ScaleM),
-				"",
-				1,
-				"L",
-				false,
-				0,
-				"",
-			)
-		}
-	}
-
-	// Floor statistics
-	if len(floor.Samples) > 0 {
-		g.pdf.Ln(pdfSpacingSmall)
-		stats := calculateFloorStats(floor.Samples)
-
-		g.pdf.SetFont("Arial", "B", pdfFontSizeNormal)
-		g.pdf.SetTextColor(0, 0, 0)
-		g.pdf.CellFormat(0, pdfSpacingLarge, "Coverage Statistics", "", 1, "L", false, 0, "")
-
-		g.pdf.SetFont("Arial", "", pdfFontSizeSmall)
-		floorMetrics := []struct {
-			label string
-			value string
-		}{
-			{"Coverage Score", fmt.Sprintf("%.0f%%", stats.CoverageScore)},
-			{"Average Signal", fmt.Sprintf("%d dBm", stats.AvgRSSI)},
-			{"Signal Range", fmt.Sprintf("%d to %d dBm", stats.MinRSSI, stats.MaxRSSI)},
-			{"Weak Spots", strconv.Itoa(stats.WeakAreas)},
-		}
-
-		for _, m := range floorMetrics {
-			g.addMetricRow(m.label, m.value)
-		}
-
-		// Channel usage summary
-		channels := getChannelUsage(floor.Samples)
-		if len(channels) > 0 {
-			g.pdf.Ln(pdfSpacingSmall)
-			g.pdf.SetFont("Arial", "B", pdfFontSizeNormal)
-			g.pdf.CellFormat(0, pdfSpacingLarge, "WiFi Channels Detected", "", 1, "L", false, 0, "")
-
-			g.pdf.SetFont("Arial", "", pdfFontSizeSmall)
-			for _, ch := range channels {
-				g.pdf.CellFormat(
-					0,
-					pdfSpacingMedium,
-					fmt.Sprintf("Channel %d: %d APs", ch.Channel, ch.Count),
-					"",
-					1,
-					"L",
-					false,
-					0,
-					"",
-				)
-			}
-		}
-	} else {
-		g.pdf.Ln(pdfSpacingSmall)
-		g.pdf.SetFont("Arial", "I", pdfFontSizeSmall)
-		g.pdf.SetTextColor(pdfColorGrayMuted, pdfColorGrayMuted, pdfColorGrayMuted)
-		g.pdf.CellFormat(0, pdfSpacingLarge, "No samples collected for this floor", "", 1, "L", false, 0, "")
-	}
-
-	// Add heatmap if requested and floor has data
-	if g.options.IncludeHeatmaps && len(floor.Samples) > 0 && floor.FloorPlan != nil {
-		g.addFloorHeatmapNote(floor)
-	}
-}
-
-// addFloorHeatmapNote adds a note about heatmap availability.
-func (g *ReportGenerator) addFloorHeatmapNote(_ *Floor) {
-	g.pdf.Ln(pdfSpacingSection)
-	g.pdf.SetFont("Arial", "I", pdfFontSizeSmall)
-	g.pdf.SetTextColor(pdfColorGrayMedium, pdfColorGrayMedium, pdfColorGrayMedium)
-	g.pdf.CellFormat(
-		0,
-		pdfSpacingMedium,
-		"Heatmap visualization available in the web interface.",
-		"",
-		1,
-		"L",
-		false,
-		0,
-		"",
-	)
-}
-
-// addRecommendations adds the recommendations section.
-func (g *ReportGenerator) addRecommendations() {
-	g.pdf.AddPage()
-	g.addSectionHeader("Recommendations")
-
-	allSamples := g.survey.GetAllSamples()
-	stats := calculateSurveyStats(allSamples)
-
-	recommendations := generateSurveyRecommendations(&stats)
-
-	if len(recommendations) == 0 {
-		g.pdf.SetFont("Arial", "I", pdfFontSizeSmall)
-		g.pdf.SetTextColor(pdfColorGrayLight, pdfColorGrayLight, pdfColorGrayLight)
-		g.pdf.CellFormat(
-			0,
-			pdfSpacingLarge,
-			"No specific recommendations - WiFi coverage meets quality standards.",
-			"",
-			1,
-			"L",
-			false,
-			0,
-			"",
-		)
-		return
-	}
-
-	g.pdf.SetFont("Arial", "", pdfFontSizeSmall)
-	g.pdf.SetTextColor(0, 0, 0)
-
-	for i, rec := range recommendations {
-		g.pdf.Ln(pdfSpacingTiny)
-		priority := getPriorityLabel(rec.Priority)
-
-		// Priority badge
-		g.pdf.SetFont("Arial", "B", pdfFontSizePriority)
-		switch rec.Priority {
-		case PriorityHigh:
-			g.pdf.SetTextColor(priorityHighColorR, priorityHighColorG, priorityHighColorB)
-		case PriorityMedium:
-			g.pdf.SetTextColor(priorityMediumColorR, priorityMediumColorG, priorityMediumColorB)
-		case PriorityLow:
-			g.pdf.SetTextColor(priorityLowColorR, priorityLowColorG, priorityLowColorB)
-		}
-		g.pdf.CellFormat(pdfSpacingPriority, pdfSpacingMedium, priority, "1", 0, "C", false, 0, "")
-
-		// Recommendation text
-		g.pdf.SetFont("Arial", "", pdfFontSizeSmall)
-		g.pdf.SetTextColor(0, 0, 0)
-		g.pdf.MultiCell(0, pdfSpacingMedium, fmt.Sprintf(" %d. %s", i+1, rec.Text), "", "L", false)
-	}
-
-	// Implementation notes
-	g.pdf.Ln(pdfSpacingSection)
-	g.pdf.SetFont("Arial", "B", pdfFontSizeNormal)
-	g.pdf.CellFormat(0, pdfSpacingLarge, "Implementation Notes", "", 1, "L", false, 0, "")
-
-	g.pdf.SetFont("Arial", "", pdfFontSizeSmall)
-	notes := []string{
-		"High priority items should be addressed within 1-2 weeks",
-		"Consider WiFi 6/6E access points for improved capacity",
-		"Verify power levels and channel assignments after changes",
-		"Re-survey affected areas after implementing changes",
-	}
-
-	for _, note := range notes {
-		g.pdf.SetTextColor(pdfColorGrayMedium, pdfColorGrayMedium, pdfColorGrayMedium)
-		g.pdf.CellFormat(pdfSpacingSmall, pdfSpacingMedium, "-", "", 0, "L", false, 0, "")
-		g.pdf.CellFormat(0, pdfSpacingMedium, note, "", 1, "L", false, 0, "")
-	}
-}
-
-// addRawDataAppendix adds the raw data appendix section.
-func (g *ReportGenerator) addRawDataAppendix() {
-	g.pdf.AddPage()
-	g.addSectionHeader("Appendix: Raw Sample Data")
-
-	allSamples := g.survey.GetAllSamples()
-	if len(allSamples) == 0 {
-		g.pdf.SetFont("Arial", "I", pdfFontSizeSmall)
-		g.pdf.CellFormat(0, pdfSpacingLarge, "No sample data collected", "", 1, "L", false, 0, "")
-		return
-	}
-
-	// Table header
-	g.pdf.SetFont("Arial", "B", pdfFontSizeTableHeader)
-	g.pdf.SetFillColor(pdfColorGrayTableBg, pdfColorGrayTableBg, pdfColorGrayTableBg)
-	headers := []struct {
-		text  string
-		width float64
-	}{
-		{"#", tableColIndexWidth},
-		{"X", tableColCoordWidth},
-		{"Y", tableColCoordWidth},
-		{"RSSI", tableColRSSIWidth},
-		{"SNR", tableColSNRWidth},
-		{"SSID", tableColSSIDWidth},
-		{"Channel", tableColChannelWidth},
-		{"Time", tableColTimeWidth},
-	}
-
-	for _, h := range headers {
-		g.pdf.CellFormat(h.width, pdfSpacingNormal, h.text, "1", 0, "C", true, 0, "")
-	}
-	g.pdf.Ln(-1)
-
-	// Table rows (limit samples per page section)
-	g.pdf.SetFont("Arial", "", pdfFontSizeTiny)
-	maxSamples := min(len(allSamples), pdfMaxSamplesInTable)
-
-	for i := range maxSamples {
-		sample := allSamples[i]
-
-		// Get first network from passive sample for display
-		rssi := "-"
-		snr := "-"
-		ssid := "-"
-		channel := "-"
-		ps := getPassiveSampleFromPoint(sample)
-		if ps != nil && len(ps.Networks) > 0 {
-			net := ps.Networks[0]
-			rssi = strconv.Itoa(net.Signal)
-			snr = strconv.Itoa(net.SNR)
-			ssid = truncateString(net.SSID, pdfSSIDMaxLength)
-			channel = strconv.Itoa(net.Channel)
-		}
-
-		g.pdf.CellFormat(tableColIndexWidth, pdfSpacingMedium, strconv.Itoa(i+1), "1", 0, "C", false, 0, "")
-		g.pdf.CellFormat(tableColCoordWidth, pdfSpacingMedium, strconv.Itoa(sample.X), "1", 0, "C", false, 0, "")
-		g.pdf.CellFormat(tableColCoordWidth, pdfSpacingMedium, strconv.Itoa(sample.Y), "1", 0, "C", false, 0, "")
-		g.pdf.CellFormat(tableColRSSIWidth, pdfSpacingMedium, rssi, "1", 0, "C", false, 0, "")
-		g.pdf.CellFormat(tableColSNRWidth, pdfSpacingMedium, snr, "1", 0, "C", false, 0, "")
-		g.pdf.CellFormat(tableColSSIDWidth, pdfSpacingMedium, ssid, "1", 0, "L", false, 0, "")
-		g.pdf.CellFormat(tableColChannelWidth, pdfSpacingMedium, channel, "1", 0, "C", false, 0, "")
-		g.pdf.CellFormat(
-			tableColTimeWidth,
-			pdfSpacingMedium,
-			sample.Timestamp.Format("15:04:05"),
-			"1",
-			0,
-			"C",
-			false,
-			0,
-			"",
-		)
-		g.pdf.Ln(-1)
-	}
-
-	if len(allSamples) > maxSamples {
-		g.pdf.Ln(pdfSpacingSmall)
-		g.pdf.SetFont("Arial", "I", pdfFontSizePriority)
-		g.pdf.CellFormat(
-			0,
-			pdfSpacingMedium,
-			fmt.Sprintf(
-				"... and %d more samples (truncated for readability)",
-				len(allSamples)-maxSamples,
-			),
-			"",
-			1,
-			"L",
-			false,
-			0,
-			"",
-		)
-	}
-}
-
-// Helper methods
-
-func (g *ReportGenerator) addSectionHeader(title string) {
-	g.pdf.SetFont("Arial", "B", pdfFontSizeSectionHead)
-	g.pdf.SetTextColor(0, 0, 0)
-	g.pdf.CellFormat(0, pdfSpacingSectionTitle, title, "", 1, "L", false, 0, "")
-
-	// Underline
-	g.pdf.SetDrawColor(pdfColorGrayLine, pdfColorGrayLine, pdfColorGrayLine)
-	g.pdf.Line(pdfLineStartX, g.pdf.GetY(), pdfLineEndX, g.pdf.GetY())
-	g.pdf.Ln(pdfSpacingSmall)
-}
-
-func (g *ReportGenerator) addStatCard(label, value, grade string) {
-	g.pdf.SetFont("Arial", "B", pdfFontSizeStatLabel)
-	g.pdf.SetTextColor(0, 0, 0)
-	g.pdf.CellFormat(0, pdfSpacingSection, label, "", 1, "L", false, 0, "")
-
-	g.pdf.SetFont("Arial", "B", pdfFontSizeStatValue)
-	gradeColor := getGradeColor(grade)
-	g.pdf.SetTextColor(gradeColor[0], gradeColor[1], gradeColor[2])
-	g.pdf.CellFormat(pdfStatValueWidth, pdfSpacingMajor, value, "", 0, "L", false, 0, "")
-
-	g.pdf.SetFont("Arial", "B", pdfFontSizeStatGrade)
-	g.pdf.CellFormat(0, pdfSpacingMajor, grade, "", 1, "L", false, 0, "")
-}
-
-func (g *ReportGenerator) addMetricRow(label, value string) {
-	g.pdf.SetTextColor(pdfColorGrayMedium, pdfColorGrayMedium, pdfColorGrayMedium)
-	g.pdf.CellFormat(pdfMetricLabelWidth, pdfSpacingMedium, label+":", "", 0, "L", false, 0, "")
-	g.pdf.SetTextColor(0, 0, 0)
-	g.pdf.CellFormat(0, pdfSpacingMedium, value, "", 1, "L", false, 0, "")
-}
-
-func (g *ReportGenerator) addDistributionBar(label string, percent float64, col []int) {
-	g.pdf.SetTextColor(pdfColorGrayMedium, pdfColorGrayMedium, pdfColorGrayMedium)
-	g.pdf.CellFormat(pdfDistLabelWidth, pdfSpacingMedium, label, "", 0, "L", false, 0, "")
-
-	// Draw bar background
-	barX := g.pdf.GetX()
-	barY := g.pdf.GetY()
-	barWidth := float64(pdfDistBarWidth)
-	barHeight := float64(pdfDistBarHeight)
-
-	g.pdf.SetFillColor(pdfDistBarBgColor, pdfDistBarBgColor, pdfDistBarBgColor)
-	g.pdf.Rect(barX, barY, barWidth, barHeight, "F")
-
-	// Draw filled portion
-	filledWidth := barWidth * (percent / percentMultiplier)
-	g.pdf.SetFillColor(col[0], col[1], col[2])
-	g.pdf.Rect(barX, barY, filledWidth, barHeight, "F")
-
-	// Percentage label
-	g.pdf.SetX(barX + barWidth + pdfDistBarGap)
-	g.pdf.SetTextColor(0, 0, 0)
-	g.pdf.CellFormat(pdfDistPercentWidth, pdfSpacingMedium, fmt.Sprintf("%.1f%%", percent), "", 1, "L", false, 0, "")
-}
-
-// SurveyStats holds calculated statistics for a survey.
-//
-//nolint:revive // Renaming to Stats would be a breaking API change
-type SurveyStats struct {
-	TotalSamples     int
-	AvgRSSI          int
-	MinRSSI          int
-	MaxRSSI          int
-	CoverageScore    float64
-	WeakAreas        int
-	DeadZones        int
-	ExcellentPercent float64
-	GoodPercent      float64
-	FairPercent      float64
-	PoorPercent      float64
-	DeadPercent      float64
-}
-
-func calculateSurveyStats(samples []*SamplePoint) SurveyStats {
-	stats := SurveyStats{
-		TotalSamples: len(samples),
-		MinRSSI:      0,
-		MaxRSSI:      minRSSIValue,
-	}
-
-	if len(samples) == 0 {
-		return stats
-	}
-
-	var sumRSSI int
-	var excellent, good, fair, poor, dead int
-
-	for _, sample := range samples {
-		ps := getPassiveSampleFromPoint(sample)
-		if ps == nil || len(ps.Networks) == 0 {
-			continue
-		}
-
-		// Use best RSSI from networks in the passive sample
-		bestRSSI := minRSSIValue
-		for _, net := range ps.Networks {
-			if net.Signal > bestRSSI {
-				bestRSSI = net.Signal
-			}
-		}
-
-		sumRSSI += bestRSSI
-		if bestRSSI > stats.MaxRSSI {
-			stats.MaxRSSI = bestRSSI
-		}
-		if bestRSSI < stats.MinRSSI || stats.MinRSSI == 0 {
-			stats.MinRSSI = bestRSSI
-		}
-
-		switch {
-		case bestRSSI > ExcellentSignal:
-			excellent++
-		case bestRSSI > GoodSignal:
-			good++
-		case bestRSSI > FairSignal:
-			fair++
-		case bestRSSI > PoorSignal:
-			poor++
-		default:
-			dead++
-		}
-	}
-
-	if len(samples) > 0 {
-		stats.AvgRSSI = sumRSSI / len(samples)
-		total := float64(len(samples))
-		stats.ExcellentPercent = float64(excellent) / total * percentMultiplier
-		stats.GoodPercent = float64(good) / total * percentMultiplier
-		stats.FairPercent = float64(fair) / total * percentMultiplier
-		stats.PoorPercent = float64(poor) / total * percentMultiplier
-		stats.DeadPercent = float64(dead) / total * percentMultiplier
-		stats.CoverageScore = stats.ExcellentPercent + stats.GoodPercent + (stats.FairPercent * fairCoverageWeight)
-		stats.WeakAreas = poor
-		stats.DeadZones = dead
-	}
-
-	return stats
-}
-
-func calculateFloorStats(samples []*SamplePoint) SurveyStats {
-	return calculateSurveyStats(samples)
-}
-
-// ChannelInfo represents WiFi channel usage information.
-type ChannelInfo struct {
-	Channel int
-	Count   int
-}
-
-func getChannelUsage(samples []*SamplePoint) []ChannelInfo {
-	channelMap := make(map[int]int)
-
-	for _, sample := range samples {
-		ps := getPassiveSampleFromPoint(sample)
-		if ps == nil {
-			continue
-		}
-		for _, net := range ps.Networks {
-			if net.Channel > 0 {
-				channelMap[net.Channel]++
-			}
-		}
-	}
-
-	channels := make([]ChannelInfo, 0, len(channelMap))
-	for ch, count := range channelMap {
-		channels = append(channels, ChannelInfo{Channel: ch, Count: count})
-	}
-
-	sort.Slice(channels, func(i, j int) bool {
-		return channels[i].Count > channels[j].Count
-	})
-
-	// Limit to top channels
-	if len(channels) > topChannelsLimit {
-		channels = channels[:topChannelsLimit]
-	}
-
-	return channels
-}
-
-// RecommendationPriority indicates the urgency of a recommendation.
-type RecommendationPriority int
-
-// Recommendation priority levels.
-const (
-	PriorityLow RecommendationPriority = iota
-	PriorityMedium
-	PriorityHigh
-)
-
-// Recommendation represents an improvement suggestion.
-type Recommendation struct {
-	Text     string
-	Priority RecommendationPriority
-}
-
-func generateSurveyRecommendations(stats *SurveyStats) []Recommendation {
-	var recommendations []Recommendation
-
-	// Coverage-based recommendations
-	switch {
-	case stats.CoverageScore < coverageScoreCritical:
-		recommendations = append(recommendations, Recommendation{
-			Text:     "Critical coverage issues detected. Consider a complete WiFi infrastructure redesign with additional access points.",
-			Priority: PriorityHigh,
-		})
-	case stats.CoverageScore < coverageScorePoor:
-		recommendations = append(recommendations, Recommendation{
-			Text:     "Poor overall coverage. Add 2-3 additional access points in strategic locations to improve connectivity.",
-			Priority: PriorityHigh,
-		})
-	case stats.CoverageScore < coverageScoreModerate:
-		recommendations = append(recommendations, Recommendation{
-			Text:     "Moderate coverage detected. Consider adding 1-2 access points to strengthen weak areas.",
-			Priority: PriorityMedium,
-		})
-	}
-
-	// Dead zone recommendations
-	if stats.DeadZones > 0 {
-		recommendations = append(recommendations, Recommendation{
-			Text: fmt.Sprintf(
-				"Found %d dead zone(s) with signal below -85 dBm. Prioritize these areas for immediate AP placement.",
-				stats.DeadZones,
-			),
-			Priority: PriorityHigh,
-		})
-	}
-
-	// Weak area recommendations
-	if stats.WeakAreas > 0 {
-		recommendations = append(recommendations, Recommendation{
-			Text: fmt.Sprintf(
-				"Found %d weak area(s) with poor signal. Consider power adjustments or additional coverage.",
-				stats.WeakAreas,
-			),
-			Priority: PriorityMedium,
-		})
-	}
-
-	// Sample density recommendations
-	if stats.TotalSamples < minSampleThreshold {
-		recommendations = append(recommendations, Recommendation{
-			Text:     "Limited sample data. Collect more samples for accurate analysis, especially in edge areas and around obstacles.",
-			Priority: PriorityLow,
-		})
-	}
-
-	// No issues found
-	if len(recommendations) == 0 && stats.CoverageScore >= coverageScoreGood {
-		recommendations = append(recommendations, Recommendation{
-			Text:     "WiFi coverage meets quality standards. Continue monitoring for any future degradation.",
-			Priority: PriorityLow,
-		})
-	}
-
-	return recommendations
-}
-
-func getPriorityLabel(p RecommendationPriority) string {
-	switch p {
-	case PriorityHigh:
-		return "HIGH"
-	case PriorityMedium:
-		return "MED"
-	case PriorityLow:
-		return "LOW"
-	}
-	return "LOW" // Default fallback for unknown priority
-}
-
-func getStatusColor(status Status) []int {
-	switch status {
-	case StatusCompleted:
-		return []int{40, 167, 69} // Green
-	case StatusInProgress:
-		return []int{0, 123, 255} // Blue
-	case StatusPaused:
-		return []int{255, 193, 7} // Yellow
-	case StatusCreated:
-		return []int{108, 117, 125} // Gray
-	}
-	return []int{108, 117, 125} // Default fallback
-}
-
-func getCoverageGrade(score float64) string {
-	switch {
-	case score >= coverageScoreGood:
-		return "A"
-	case score >= coverageGradeB:
-		return "B"
-	case score >= coverageGradeC:
-		return "C"
-	case score >= coverageGradeD:
-		return "D"
-	default:
-		return "F"
-	}
-}
-
-func getGradeColor(grade string) []int {
-	switch grade {
-	case "A":
-		return []int{40, 167, 69}
-	case "B":
-		return []int{144, 238, 144}
-	case "C":
-		return []int{255, 193, 7}
-	case "D":
-		return []int{255, 128, 0}
-	default:
-		return []int{220, 53, 69}
-	}
-}
-
-func truncateString(s string, maxLen int) string {
-	if len(s) <= maxLen {
-		return s
-	}
-	return s[:maxLen-3] + "..."
-}
-
-// getPassiveSampleFromPoint extracts PassiveSample data from a SamplePoint.
-// Returns nil if the SampleData is not a PassiveSample.
-func getPassiveSampleFromPoint(sp *SamplePoint) *PassiveSample {
-	if sp == nil || sp.SampleData == nil {
-		return nil
-	}
-	switch data := sp.SampleData.(type) {
-	case *PassiveSample:
-		return data
-	case PassiveSample:
-		return &data
-	default:
-		return nil
-	}
 }
 
 // GenerateReport creates a PDF report for the specified survey.
